# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_all-in-one_internal_http.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.CheckRequest:
    properties:
      bucket_key:
        type: string
      target_key:
        type: string
    type: object
  model.CheckResponse:
    properties:
      allowed:
        type: boolean
      limit:
        type: integer
      remaining:
        type: integer
      retry_after_seconds:
        type: integer
    type: object
...
  model.Kind:
    enum:
    - throttle
    - daily_quota
    type: string
    x-enum-varnames:
    - KindThrottle
    - KindDailyQuota
...
  model.Scope:
    enum:
    - ip
    - user
    - global
    type: string
    x-enum-varnames:
    - ScopeIP
    - ScopeUser
    - ScopeGlobal
...
  model.Target:
    properties:
      app:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      is_external:
        type: boolean
      key:
        type: string
      kind:
        $ref: '#/definitions/model.Kind'
      limit_count:
        type: integer
      method:
        type: string
      name:
        type: string
      path:
        type: string
      scope:
        $ref: '#/definitions/model.Scope'
      updated_at:
        type: string
      updated_by:
        type: string
      window_unit:
        $ref: '#/definitions/model.WindowUnit'
      window_value:
        type: integer
    type: object
  model.TargetPatch:
    properties:
      enabled:
        type: boolean
      limit_count:
        type: integer
      window_unit:
        $ref: '#/definitions/model.WindowUnit'
      window_value:
        type: integer
    type: object
...
  model.WindowUnit:
    enum:
    - second
    - minute
    - hour
    - day
    type: string
    x-enum-varnames:
    - WindowSecond
    - WindowMinute
    - WindowHour
    - WindowDay
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: All-in-one is a multi-functional apps with the main goal for learning
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: All-in-one API
  version: "1.0"
paths:
...
  /ratelimit/check:
    post:
      consumes:
      - application/json
      description: Evaluate one external target against a caller-supplied bucket key.
        Authenticated by X-API-Key (an app token), not a user session. A rejection
        is reported as 200 with allowed=false (not 429) so the caller can tell "your
        user is limited" from "aio limited you". Fails open on an internal counter
        error.
      parameters:
      - description: App token
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Target and bucket key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Decision
          schema:
            allOf:
            - $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.CheckResponse'
              type: object
        "400":
          description: Invalid request or bucket key does not match scope
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "401":
          description: Missing or invalid app token
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "403":
          description: Target key outside token scope
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "404":
          description: Unknown external target
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "503":
          description: External rate limiting disabled
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
      summary: Check a rate limit (external apps)
      tags:
      - rate-limiting
  /ratelimit/targets:
    get:
      description: List every rate-limited target, merging its code-defined identity
        (scope/kind/route) with its current effective rule (admin-only)
      produces:
      - application/json
      responses:
        "200":
          description: List of targets
          schema:
            allOf:
            - $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Target'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "403":
          description: Forbidden (not an admin)
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
      security:
      - BearerAuth: []
        DirectAuth: []
      summary: List rate limit targets
      tags:
      - rate-limiting
  /ratelimit/targets/{key}:
    delete:
      description: Delete a DB-backed external target by key (admin-only). Internal
        (Registry) targets cannot be deleted.
      parameters:
      - description: Target key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "400":
          description: Not an external target
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "403":
          description: Forbidden (not an admin)
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "404":
          description: Unknown target
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
      security:
      - BearerAuth: []
        DirectAuth: []
      summary: Delete an external rate limit target
      tags:
      - rate-limiting
    patch:
      consumes:
      - application/json
      description: Partially update a target's enabled/limit/window (admin-only).
        Omitted fields are left unchanged (pointer semantics).
      parameters:
      - description: Target key
        in: path
        name: key
        required: true
        type: string
      - description: Fields to update
        in: body
        name: patch
        required: true
        schema:
          $ref: '#/definitions/model.TargetPatch'
      produces:
      - application/json
      responses:
        "200":
          description: Updated target
          schema:
            allOf:
            - $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Target'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "403":
          description: Forbidden (not an admin)
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "404":
          description: Unknown target
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
      security:
      - BearerAuth: []
        DirectAuth: []
      summary: Update a rate limit target's rule
      tags:
      - rate-limiting
  /ratelimit/targets/{key}/reset:
    post:
      description: Clears today's daily-quota counters for a target (admin-only).
        Does not clear an in-memory throttle bucket (per-process state).
      parameters:
      - description: Target key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reset acknowledged
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "403":
          description: Forbidden (not an admin)
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "404":
          description: Unknown target
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
      security:
      - BearerAuth: []
        DirectAuth: []
      summary: Reset a target's counters
      tags:
      - rate-limiting
  /ratelimit/targets/{key}/reset-defaults:
    post:
      description: Overwrites a target's enabled/limit/window back to its Registry
        defaults, clearing any prior admin edit (admin-only)
      parameters:
      - description: Target key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Target reset to defaults
          schema:
            allOf:
            - $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Target'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "403":
          description: Forbidden (not an admin)
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "404":
          description: Unknown target
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
      security:
      - BearerAuth: []
        DirectAuth: []
      summary: Reset a target's rule to its code-defined defaults
      tags:
      - rate-limiting
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  DirectAuth:
    description: Direct authentication bypass for testing (only works when DirectAuthEnabled
      is true in config). Provide username directly.
    in: header
    name: x-direct-auth-username
    type: apiKey
swagger: "2.0"